Fix a translation error
The Chinese README states "每日触达数亿用户", which in English translates to "reaches hundreds of millions of users daily", but the English README states it reaches "tens of billions" of users (there are only 7 billion people on the planet). This should read "hundreds of millions".

I also changed "concerning" to "reaching" as it's a more natural word for this context.